requirements.txt
fpdf
Pillow
docx2pdf
python-pptx
reportlab